# bank-employees-salaries-api/resources/openapi.yaml
openapi: 3.0.1
info:
  title: API для работы с данными посещаемости, отделами и сотрудниками
  version: 1.0.0
  description: |
    Система формирования отчетов для AES.
    Включает компоненты для работы с моделями и DTO.
servers:
  - url: http://localhost:8080/api
    description: Локальный сервер разработки
  - url: https://api.example.com/v1
    description: Продуктивный сервер
...
paths:
...
  /employee:
    post:
      tags: [ Employee ]
      summary: Добавить нового сотрудника
      operationId: createEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'EmployeeDto.yaml#/components/schemas/Employee'
      responses:
        '201':
          description: Сотрудник создан
          content:
            application/json:
              schema:
                $ref: 'EmployeeDto.yaml#/components/schemas/EmployeeDto'
        '400':
          description: Неверные данные
        '500':
          description: Внутренняя ошибка сервера

    get:
      tags: [ Employee ]
      summary: Получить список всех сотрудников
      operationId: getAllEmployees
      responses:
        '200':
          description: Список сотрудников
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'EmployeeDto.yaml#/components/schemas/EmployeeDto'
        '500':
          description: Внутренняя ошибка сервера

    delete:
      tags: [ Employee ]
      summary: Удалить всех сотрудников
      operationId: deleteAllEmployees
      responses:
        '204':
          description: Все сотрудники удалены
        '500':
          description: Внутренняя ошибка сервера
  /employee/{id}:
    get:
      tags: [ Employee ]
      summary: Получить сотрудника по ID
      operationId: getEmployeeById
      parameters:
        - name: id
          in: path
          required: true
          description: ID сотрудника
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о сотруднике
          content:
            application/json:
              schema:
                $ref: 'EmployeeDto.yaml#/components/schemas/EmployeeDto'
        '404':
          description: Сотрудник не найден
        '500':
          description: Внутренняя ошибка сервера

    patch:
      tags: [ Employee ]
      summary: Обновить информацию о сотруднике
      operationId: updateEmployee
      parameters:
        - name: id
          in: path
          required: true
          description: ID сотрудника
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'EmployeeDto.yaml#/components/schemas/Employee'
      responses:
        '200':
          description: Информация обновлена
          content:
            application/json:
              schema:
                $ref: 'EmployeeDto.yaml#/components/schemas/EmployeeDto'
        '400':
          description: Неверные данные
        '404':
          description: Сотрудник не найден
        '500':
          description: Внутренняя ошибка сервера

    delete:
      tags: [ Employee ]
      summary: Удалить сотрудника по ID
      operationId: deleteEmployeeById
      parameters:
        - name: id
          in: path
          required: true
          description: ID сотрудника
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Сотрудник удален
        '404':
          description: Сотрудник не найден
        '500':
          description: Внутренняя ошибка сервера

  # ========== Grade Endpoints ==========
  /grade:
    post:
      tags: [ Grade ]
      summary: Добавление нового грейда
      operationId: createGrade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'GradeDto.yaml#/components/schemas/Grade'
      responses:
        '201':
          description: Грейд создан
          content:
            application/json:
              schema:
                $ref: 'GradeDto.yaml#/components/schemas/GradeDto'
        '400':
          description: Неверные данные
        '500':
          description: Внутренняя ошибка сервера

    get:
      tags: [ Grade ]
      summary: Получить список всех грейдов
      operationId: getAllGrades
      responses:
        '200':
          description: Список грейдов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'GradeDto.yaml#/components/schemas/GradeDto'
        '500':
          description: Внутренняя ошибка сервера

    delete:
      tags: [ Grade ]
      summary: Удалить все грейды
      operationId: deleteAllGrades
      responses:
        '204':
          description: Все грейды удалены
        '500':
          description: Внутренняя ошибка сервера
  /grade/{id}:
    get:
      tags: [ Grade ]
      summary: Получить грейд по ID
      operationId: getGradeById
      parameters:
        - name: id
          in: path
          required: true
          description: ID грейда
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о грейде
          content:
            application/json:
              schema:
                $ref: 'GradeDto.yaml#/components/schemas/GradeDto'
        '404':
          description: Грейд не найден
        '500':
          description: Внутренняя ошибка сервера

    patch:
      tags: [ Grade ]
      summary: Обновить информацию о грейде
      operationId: updateGrade
      parameters:
        - name: id
          in: path
          required: true
          description: ID грейда
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'GradeDto.yaml#/components/schemas/Grade'
      responses:
        '200':
          description: Информация обновлена
          content:
            application/json:
              schema:
                $ref: 'GradeDto.yaml#/components/schemas/GradeDto'
        '400':
          description: Неверные данные
        '404':
          description: Грейд не найден
        '500':
          description: Внутренняя ошибка сервера

    delete:
      tags: [ Grade ]
      summary: Удалить грейд по ID
      operationId: deleteGradeById
      parameters:
        - name: id
          in: path
          required: true
          description: ID сотрудника
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Грейд удален
        '404':
          description: Сотрудник не найден
        '500':
          description: Внутренняя ошибка сервера

  # ========== Kpis Endpoints ==========
  /kpis:
    post:
      operationId: addKpi
      summary: Добавление нового KPI
      description: Создает новый показатель KPI в системе
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'KpiDto.yaml#/components/schemas/Kpi'
      responses:
        '200':
          description: KPI успешно создан
          content:
            application/json:
              schema:
                $ref: 'KpiDto.yaml#/components/schemas/KpiDto'
        '400':
          description: Неверные входные данные
        '500':
          description: Внутренняя ошибка сервера
      tags:
        - KPI

    get:
      operationId: getKpis
      summary: Получение списка всех KPI
      description: Возвращает список всех показателей KPI из системы
      responses:
        '200':
          description: Список KPI успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'KpiDto.yaml#/components/schemas/KpiDto'
        '500':
          description: Внутренняя ошибка сервера
      tags:
        - KPI

    delete:
      operationId: deleteKpis
      summary: Удаление всех KPI
      description: Удаляет все показатели KPI из системы
      responses:
        '204':
          description: Все KPI успешно удалены
        '500':
          description: Внутренняя ошибка сервера
      tags:
        - KPI
  /kpis/{id}:
    get:
      operationId: getKpiById
      summary: Получение KPI по ID
      description: Возвращает показатель KPI по указанному идентификатору
      parameters:
        - name: id
          in: path
          description: ID KPI
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '200':
          description: KPI успешно получен
          content:
            application/json:
              schema:
                $ref: 'KpiDto.yaml#/components/schemas/KpiDto'
        '404':
          description: KPI не найден
        '400':
          description: Неверный ID (отрицательное число)
        '500':
          description: Внутренняя ошибка сервера
      tags:
        - KPI

    patch:
      operationId: updateKpi
      summary: Обновление KPI
      description: Обновляет существующий показатель KPI
      parameters:
        - name: id
          in: path
          description: ID KPI для обновления
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'KpiDto.yaml#/components/schemas/Kpi'
      responses:
        '200':
          description: KPI успешно обновлен
          content:
            application/json:
              schema:
                $ref: 'KpiDto.yaml#/components/schemas/KpiDto'
        '400':
          description: Неверные входные данные
        '404':
          description: KPI не найден
        '500':
          description: Внутренняя ошибка сервера
      tags:
        - KPI

    delete:
      operationId: deleteKpiById
      summary: Удаление KPI по ID
      description: Удаляет показатель KPI по указанному идентификатору
      parameters:
        - name: id
          in: path
          description: ID KPI для удаления
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '204':
          description: KPI успешно удален
        '400':
          description: Неверный ID (отрицательное число)
        '404':
          description: KPI не найден
        '500':
          description: Внутренняя ошибка сервера
      tags:
        - KPI

  # ========== Position Endpoints ==========
...